Move gRPC error intercetpors to api/utils/grpc/interceptors.
